pyproject.toml
[project]
name = "tiny-dec"
version = "0.1.0"
description = "An education-purpose tiny decompiler"
authors = [
    {name = "Zhuo Zhang",email = "zz@cs.columbia.edu"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "pwntools (>=4.15.0,<5.0.0)"
]

[project.scripts]
tiny-dec = "tiny_dec.cli:main"

[dependency-groups]
dev = [
    "pytest (>=8.4.0,<9.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "ruff (>=0.15.0,<0.16.0)"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
mypy_path = ["stubs"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
